backend/requirements.txt
fastapi>=0.100.0
hypercorn>=0.14.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
openai>=1.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
playwright>=1.40.0
